Resolved build errors

+++ src/SFA.DAS.EmploymentVerification.Database/Tables/CheckType.sql
@@ -1,6 +1,6 @@
 CREATE TABLE [dbo].[CheckType]
 (
-  CheckTypeId INT NOT NULL Identity(1,1) Constraint PK_CheckType PRIMARY KEY,
+  CheckTypeId SMALLINT NOT NULL Identity(1,1) Constraint PK_CheckType PRIMARY KEY,
   CheckType VARCHAR(20)
 );
 GO
