+++ src/SFA.DAS.EmploymentVerification.Database/Tables/CommitmentStatus.sql
@@ -1,6 +1,6 @@
 CREATE TABLE [dbo].[CommitmentStatus]
 (
-  CommitmentStatusId INT NOT NULL Identity(1,1) Constraint PK_CommitmentStatus PRIMARY KEY,
+  CommitmentStatusId SMALLINT NOT NULL Identity(1,1) Constraint PK_CommitmentStatus PRIMARY KEY,
   CommitmentStatus VARCHAR(20)
 );
 GO
